refactor dpm pm quality review action routes
